scripts/previewImageWebpConverter/requirements.txt
boto3==1.34.157
botocore==1.34.157
cairocffi==1.7.1
CairoSVG==2.7.1
cffi==1.17.0
cssselect2==0.7.0
defusedxml==0.7.1
greenlet==3.0.3
jmespath==1.0.1
pillow==10.4.0
playwright==1.45.1
pycparser==2.22
pyee==11.1.0
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
tinycss2==1.3.0
typing_extensions==4.12.2
urllib3==2.2.2
webencodings==0.5.1
